reduce unnecessary rendering from state changes
