+++ frontend/public/redux.js
@@ -1,10 +1,16 @@
-import { applyMiddleware, combineReducers, createStore } from 'redux';
+import { applyMiddleware, combineReducers, createStore, compose } from 'redux';
 import { reducer as formReducer } from 'redux-form';
 
 import { featureReducer, featureReducerName } from './features';
 import { monitoringReducer, monitoringReducerName } from './monitoring';
 import k8sReducers from './module/k8s/k8s-reducers';
 import UIReducers from './ui/ui-reducers';
+
+const composeEnhancers =
+  // eslint-disable-next-line no-undef
+  (process.env.NODE_ENV !== 'production' &&
+    window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) ||
+  compose;
 
 /**
  * This is the entirety of the `redux-thunk` library.
@@ -31,7 +37,7 @@ const reducers = combineReducers({
   [monitoringReducerName]: monitoringReducer,
 });
 
-const store = createStore(reducers, {}, applyMiddleware(thunk));
+const store = createStore(reducers, {}, composeEnhancers(applyMiddleware(thunk)));
 export default store;
 
 // eslint-disable-next-line no-undef
